[BE] feat : 이의제기 수정, 삭제 기능
글쓴이가 동일하고, 경찰이 답을 하지 않은 경우에만 수정 및 삭제 가능

+++ kickcap-BE/src/main/java/com/ssafy/kickcap/objection/repository/ObjectionRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.kickcap.objection.repository;
 import com.ssafy.kickcap.objection.entity.Objection;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ObjectionRepository extends JpaRepository<Objection, Integer> {
+public interface ObjectionRepository extends JpaRepository<Objection, Long> {
 
 }
